# specs/000-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Manager API
  version: 0.3.0
  description: Contrato inicial do MVP. Endpoints autenticados usam sessão em cookie.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
...
paths:
...
  /admin/tenants:
    get:
      tags: [Administration]
      operationId: listTenants
      summary: Listar tenants para administração
      x-requirements: [US-001, FR-002, FR-003]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Página de tenants visível ao PlatformAdmin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantPage' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Administration]
      operationId: createTenant
      summary: Criar tenant e proprietário
      x-requirements: [US-001, US-008, FR-002, FR-003, FR-019, FR-025, BR-013, BR-014]
      parameters:
        - $ref: '#/components/parameters/AntiforgeryToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTenantRequest' }
      responses:
        '201':
          description: Tenant, proprietário convidado e link manual retornado uma única vez
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantProvisioningResult' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-waai-session
  parameters:
    AntiforgeryToken:
      in: header
      name: X-CSRF-TOKEN
      required: true
      schema: { type: string, minLength: 16 }
...
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
    PageLimit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    CreateTenantRequest:
      type: object
      required: [name, slug, ownerEmail, officialApiLineCount, qrCodeLineCount]
      properties:
        name: { type: string, minLength: 1, maxLength: 150 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        ownerEmail: { type: string, format: email }
        officialApiLineCount: { type: integer, minimum: 0, default: 0 }
        qrCodeLineCount: { type: integer, minimum: 0, default: 0 }
...
    TenantSummary:
      type: object
      required: [id, name, slug, status, version, officialApiLineCount, qrCodeLineCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [Active, Suspended] }
        version: { type: integer, minimum: 0 }
        officialApiLineCount: { type: integer, minimum: 0 }
        qrCodeLineCount: { type: integer, minimum: 0 }
    TenantProvisioningResult:
      type: object
      required: [tenant, ownerInvitation]
      properties:
        tenant: { $ref: '#/components/schemas/TenantSummary' }
        ownerInvitation: { $ref: '#/components/schemas/InvitationIssued' }
    TenantPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TenantSummary' }
        nextCursor: { type: [string, 'null'] }
...
    InvitationIssued:
      type: object
      required: [membershipId, activationUrl, expiresAt]
      properties:
        membershipId: { type: string, format: uuid }
        activationUrl:
          type: string
          format: uri-reference
          description: Contém token em claro e é retornada uma única vez; nunca persistir ou registrar
        expiresAt: { type: string, format: date-time }
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
security:
  - sessionCookie: []